requirements.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
pyjwt==2.8.1
cryptography==41.0.7
passlib==1.7.4
bcrypt==4.1.1

# OpenTelemetry (Observability)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0

# Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
isort==5.13.2
mypy==1.7.1
flake8==6.1.0
